Add textarea demo to the showcase
