+++ examples/showcase/Cargo.toml
@@ -14,3 +14,4 @@ game_of_life = { path = "sub/game_of_life" }
 large_table = { path = "sub/large_table" }
 mount_point = { path = "sub/mount_point" }
 npm_and_rest = { path = "sub/npm_and_rest" }
+textarea = { path = "sub/textarea" }
