+++ examples/showcase/sub/textarea/Cargo.toml
@@ -4,4 +4,4 @@ version = "0.1.0"
 authors = ["Andrew Straw <strawman@astraw.com>"]
 
 [dependencies]
-yew = { path = "../.." }
+yew = { path = "../../../.." }
