+++ examples/showcase/sub/textarea/src/main.rs
@@ -1,68 +1,12 @@
-#[macro_use]
 extern crate yew;
+extern crate textarea;
 
 use yew::prelude::*;
-
-struct Context {
-}
-
-struct Model {
-    value: String,
-}
-
-enum Msg {
-    GotInput(String),
-    Clicked,
-}
-
-impl Component<Context> for Model {
-    type Msg = Msg;
-    type Properties = ();
-
-    fn create(_: Self::Properties, _: &mut Env<Context, Self>) -> Self {
-        Model {
-            value: "".into(),
-        }
-    }
-
-    fn update(&mut self, msg: Self::Msg, _: &mut Env<Context, Self>) -> ShouldRender {
-        match msg {
-            Msg::GotInput(new_value) => {
-                self.value = new_value;
-            }
-            Msg::Clicked => {
-                self.value = "blah blah blah".to_string();
-            }
-        }
-        true
-    }
-}
-
-impl Renderable<Context, Model> for Model {
-    fn view(&self) -> Html<Context, Self> {
-        html! {
-            <div>
-                <div>
-                    <textarea rows=5,
-                        value=&self.value,
-                        oninput=|e: InputData| Msg::GotInput(e.value),
-                        placeholder="placeholder",>
-                    </textarea>
-                     <button onclick=|_| Msg::Clicked,>{ "change value" }</button>
-                </div>
-                <div>
-                    {&self.value}
-                </div>
-            </div>
-        }
-    }
-}
+use textarea::Model;
 
 fn main() {
     yew::initialize();
-    let context = Context {
-    };
-    let app: App<_, Model> = App::new(context);
+    let app: App<_, Model> = App::new(());
     app.mount_to_body();
     yew::run_loop();
 }
